requirements.txt
discord==2.0.0
discord.py==2.0.1
psycopg2==2.9.3
python-dotenv==0.21.0
requests==2.28.1